refactor(api): consolidate REST API into RVNKCore LoreController (#352)
- Remove LoreApiInitializer and LoreApiServlet (routing moved to RVNKCore)
- Add LoreApiEndpointImpl implementing ILoreApiService
- Remove jakarta.servlet-api dependency (provided by RVNKCore)
- Add finalName to pom.xml for versionless JAR output
- Update CLAUDE.md to reflect new API architecture
