+++ pom.xml
@@ -108,13 +108,7 @@
            <version>1.0-SNAPSHOT</version>
            <scope>provided</scope>
        </dependency>
-       <!-- Jakarta Servlet API (provided by RVNKCore's Jetty server at runtime) -->
-       <dependency>
-           <groupId>jakarta.servlet</groupId>
-           <artifactId>jakarta.servlet-api</artifactId>
-           <version>6.0.0</version>
-           <scope>provided</scope>
-       </dependency>
+       <!-- Jakarta Servlet API removed — API routing handled by RVNKCore's LoreController -->
        <!-- PlaceholderAPI for placeholder integration -->
        <dependency>
            <groupId>me.clip</groupId>
@@ -196,6 +190,7 @@
        </dependency>
    </dependencies>
    <build>
+       <finalName>${project.artifactId}</finalName>
        <resources>
            <resource>
                <directory>src/main/resources</directory>
